refactor: remove consol.log and restructure components
