+++ FE/src/app/home-page/search-area/search-area.component.css
