+++ FE/src/app/home-page/top-bar/top-bar.component.ts
