+++ FE/src/app/reloadable-area/loader/loader.component.html
